Update and rename webApScrape.py to webAppScrape.py
